+++ packages/adminui-vue/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0]
+### Changed
+ - vue updated 2.5.17 -> 2.6.6
+ - elemet-ui updated 2.4.9 -> 2.5.4
+
 ## [1.0.44]
 ### Added
 - watch for prop value changed in all controls